mega commit
mostly made by cryfact

updated curl/zlib, implement blt_dll_version(), improved signature searching

changes by me:
removed libssh2 dep from libcurl

+++ src/util/files.cpp
@@ -86,7 +86,7 @@ const char *IOException::exceptionName() const
 	}
 
 	void EnsurePathWritable(const std::string& path){
-		int finalSlash = path.find_last_of('/');
+		size_t finalSlash = path.find_last_of('/');
 		std::string finalPath = path.substr(0, finalSlash);
 		if (DirectoryExists(finalPath)) return;
 		CreateDirectoryPath(finalPath);
